create catedratico

+++ database/migrations/2022_10_06_075940_create_professors_table.php
@@ -15,7 +15,7 @@ class CreateProfessorsTable extends Migration
     {
         Schema::create('professor', function (Blueprint $table) {
             $table->id();
-            $table->integer('dpi');
+            $table->bigInteger('dpi');
             $table->string('name', 50);
             $table->string('last_name', 50);
  /**pendiente agregar relaciòn con escuela */
